feat: adds forward-proxy
