+++ web/version.go
@@ -1,0 +1,11 @@
+package web
+
+import (
+	"fmt"
+	"net/http"
+)
+
+func (h *handler) version(w http.ResponseWriter, r *http.Request) {
+	w.Header().Add("Content-Type", "text/html")
+	fmt.Fprintf(w, "<pre>%v</pre>", h.config.Version)
+}
